add feign client to get de price by the code

+++ src/main/java/org/globant/orderservice/model/Order.java
@@ -22,7 +22,7 @@ public class Order {
     private int id;
     private String ciUser;
     private String Address;
-    private Double total;
+    private Float total;
     @Column(nullable = false, updatable = false)
     @CreationTimestamp()
     private Date dateCreation;
